sps_tool/requirements.txt
flask>=3.0
python-docx>=1.1
openpyxl>=3.1
pandas>=2.0
anthropic>=0.30
python-dotenv>=1.0
python-dateutil>=2.9
